Move location objects to dedicated module
