+++ labonneboite/common/search.py
@@ -10,7 +10,6 @@ import elasticsearch.exceptions
 from slugify import slugify
 from backports.functools_lru_cache import lru_cache
 
-from labonneboite.common import geocoding
 from labonneboite.common import mapping as mapping_util
 from labonneboite.common import sorting
 from labonneboite.common import autocomplete
@@ -42,43 +41,6 @@ KEY_TO_LABEL_DISTANCES = {
 FILTERS = ['naf', 'headcount', 'flag_alternance', 'distance']
 DISTANCE_FILTER_MAX = 3000
 
-
-
-class Location(object):
-    def __init__(self, latitude, longitude):
-        self.latitude = latitude
-        self.longitude = longitude
-
-    def __repr__(self):
-        return "lon: {} - lat: {}".format(self.longitude, self.latitude)
-
-
-class CityLocation(object):
-
-    def __init__(self, zipcode, slug=''):
-        self.zipcode = zipcode
-
-        # Location attribute may be None if slug/zipcode combination is incorrect
-        self.location = None
-        self.slug = slug
-        self.name = slug.replace('-', ' ').capitalize()
-
-        city = geocoding.get_city_by_zipcode(self.zipcode, slug=slug)
-        if not city:
-            logger.debug("unable to retrieve a city for zipcode `%s` and slug `%s`", self.zipcode, self.slug)
-        else:
-            coordinates = city['coords']
-            self.location = Location(coordinates['lat'], coordinates['lon'])
-            self.slug = city['slug']
-            self.name = city['name']
-
-    @property
-    def is_location_correct(self):
-        return self.location is not None
-
-    @property
-    def full_name(self):
-        return '%s (%s)' % (self.name, self.zipcode)
 
 
 class Fetcher(object):
